+++ neo/Prompt/Utils.py
@@ -147,6 +147,7 @@ def attr_obj_to_tx_attr(obj):
 def parse_param(p, wallet=None, ignore_int=False, prefer_hex=True):
 
     # first, we'll try to parse an array
+
     try:
         items = eval(p, {"__builtins__": {}}, {})
         if len(items) > 0 and type(items) is list:
@@ -169,10 +170,17 @@ def parse_param(p, wallet=None, ignore_int=False, prefer_hex=True):
             pass
 
     try:
-        val = eval(p, {"__builtins__": {}}, {})
-
+        val = eval(p, {"__builtins__": {'bytearray': bytearray, 'bytes': bytes}}, {})
         if type(val) is bytearray:
-            return val.hex()
+            return val
+        elif type(val) is bytes:
+            # try to unhex
+            try:
+                val = binascii.unhexlify(val)
+            except Exception as e:
+                pass
+            # now it should be unhexxed no matter what, and we can hex it
+            return val.hex().encode('utf-8')
 
         return val
     except Exception as e:
@@ -225,12 +233,11 @@ def lookup_addr_str(wallet, addr):
 
 
 def parse_hold_vins(results):
-    print("results!!! %s " % results)
 
     holds = results[0].GetByteArray()
     holdlen = len(holds)
     numholds = int(holdlen / 33)
-    print("holds, holdlen, numholds %s %s " % (holds, numholds))
+
     vins = []
     for i in range(0, numholds):
         hstart = i * 33
@@ -239,11 +246,8 @@ def parse_hold_vins(results):
 
         vin_index = item[0]
         vin_tx_id = UInt256(data=item[1:])
-        print("VIN INDEX, VIN TX ID: %s %s" % (vin_index, vin_tx_id))
 
         t_input = TransactionInput(prevHash=vin_tx_id, prevIndex=vin_index)
-
-        print("found tinput: %s " % json.dumps(t_input.ToJson(), indent=4))
 
         vins.append(t_input)